CardGameDaft2
Worked on the card game a bit, trying to figure out functionallity through the hud and get the numbers and arrays i built before. I also created a main menu folder with a level and widget for the hud in the main menu, Gareth, everytime you try to play any level it comes up an error in your endlessrunner BP no clue what it is mate!