CardGame/Cleanup
Worked more on the card game and main menu, blocking out the main menu animation and new places for the buttons. Card game I fixed by changing the value to be 2 different ones for the player and computer when drawing which fixed a few issues regarding it not choosing the right suit when resolving the game.